cpuallocator: add CPU priority levels
Implement three levels of CPU priority (high, normal, low) instead of
binary hi/lo flagging previously used. The aim is to have a simple but a
bit more versatile scheme for evaluating CPU priorities. This scheme
could be further elaborated in the future although it might be hard to
come up with one number that could objectively represent the "priority"
of a CPU in a general case.

The patch tries to preserve the existing behavior closely. There is one
notable difference though: after this patch the existing policies (that
utilize cpuallocator) favor "normal" priority cores over "low" priority
cores when allocating cpus for non-prioritized workloads. This can have
a practical effect, especially if the sockets (CPU packages) in a system
are configured non-uniformly, e.g. so that one socket only has "normal"
cores whereas some others have only "high" and "low" priority cores. In
this case the cpuallocator would try to pack all non-prioritized
workloads on the core(s) that have CPUs with matching priority.
Currently none of the policies utilize the possibility to allocate "low"
priority cpus.